completes as much of lab as I can get to tonight; chatroom is partially functional; needs more work
