+++ package.json
@@ -4,10 +4,8 @@
   "description": "![cf](https://i.imgur.com/7v5ASc8.png) Lab 06: TCP Chat Server ======",
   "main": "index.js",
   "scripts": {
-    "test": "eslint . && jest --coverage",
-    "testWatch": "jest --coverage --watchAll",
-    "test-nolint": "jest --coverage",
-    "lint": "eslint ."
+    "start": "nodemon index.js",
+    "test": "echo \"Error, no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
