+++ src/lib/chatroom.js
@@ -1,10 +1,16 @@
 'use strict';
 
-import '../events/events';
-import 'net';
-import './logger';
-import './server';
+const EventEmitter = require('events');
+
+const net = require('net');
+const logger = require('./logger');
+const User = require('../models/users');
+
+const PORT = process.env.PORT || 3000;
+
+const server = net.createServer();
 const event = new EventEmitter();
+const clientPool = {};
 
 const parseData = (buffer) => {
   let text = buffer.toString().trim();
@@ -22,6 +28,36 @@ const parseData = (buffer) => {
 
 const dispatchAction = (user, buffer) => {
   const entry = parseData(buffer);
-  console.log(entry, 'This is the ENTRY');
+  console.log(entry, 'THIS IS THE ENTRY');
   if (entry) event.emit(entry.command, entry, user);
 };
+
+server.on('connection', (socket) => {
+  const user = new User (socket);
+  socket.write(`Welcome to my chatroom ${user.nickname}!\n`);
+  clientPool[user._id] = user;
+  logger.log(logger.INFO, `A new user named ${user.nickname} has entered the chatroom`);  
+  socket.on('data', (buffer) => {
+    console.log(buffer);
+    dispatchAction(user, buffer);
+  });
+});
+
+server.on('error', (e) => {
+  if (e.code === 'EADDRINUSE') {
+    console.log('Address in use, retrying...');
+    setTimeout(() => {
+      server.close();
+      server.listen(PORT, HOST);
+    }, 3000);
+  }
+});
+
+server.on('close_server', function() {
+  console.log('Server Closed');
+  process.exit();
+});
+
+server.listen(PORT, () => {
+  logger.log(logger.INFO, `Server is listening on Port: ${PORT}`);
+});
